+++ components/tab-bar/src/vwc-tab.ts
@@ -30,11 +30,7 @@ export class VWCTab extends MWCTab {
 	trailingIcon = false;
 
 	protected renderIcon(): TemplateResult {
-		return html`<vwc-icon
-			class="vvd-icon"
-			size="small"
-			type="${this.icon}"
-		></vwc-icon>`;
+		return html`<vwc-icon type="${this.icon}"></vwc-icon>`;
 	}
 
 	// ! copy & paste code from original mwc tab