+++ components/list/CHANGELOG.md
@@ -1,3 +1,46 @@
+# [2.12.0](https://github.com/vonage/vivid/compare/v2.11.0...v2.12.0) (2021-06-14)
+
+
+
+# [2.11.0](https://github.com/vonage/vivid/compare/v2.10.0...v2.11.0) (2021-06-07)
+
+
+
+# [2.10.0](https://github.com/vonage/vivid/compare/v2.9.1...v2.10.0) (2021-06-04)
+
+
+
+## [2.9.1](https://github.com/vonage/vivid/compare/v2.9.0...v2.9.1) (2021-05-27)
+
+
+
+# [2.9.0](https://github.com/vonage/vivid/compare/v2.8.0...v2.9.0) (2021-05-24)
+
+
+### Bug Fixes
+
+* **vwc-list-item:** issue 838 fixing list items bugs ([#854](https://github.com/vonage/vivid/issues/854)) ([e8ed53b](https://github.com/vonage/vivid/commit/e8ed53b762308a466b40cafa0f5a780f5d35537a)), closes [#838](https://github.com/vonage/vivid/issues/838) [#838](https://github.com/vonage/vivid/issues/838)
+
+
+
+# [2.8.0](https://github.com/vonage/vivid/compare/v2.7.0...v2.8.0) (2021-05-20)
+
+
+
+# [2.7.0](https://github.com/vonage/vivid/compare/v2.6.2...v2.7.0) (2021-05-17)
+
+
+### Bug Fixes
+
+* **vwc-check-list-item:** fixing the layout of the radio and checkbox elements ([#839](https://github.com/vonage/vivid/issues/839)) ([a8eea1e](https://github.com/vonage/vivid/commit/a8eea1ed55523a237f56c9b0ac50b3d0627ab158))
+
+
+### Features
+
+* **snackbar:** iss 783 viv 402 snackbar ([#808](https://github.com/vonage/vivid/issues/808)) ([25c750e](https://github.com/vonage/vivid/commit/25c750e99253a4f00e9765b30a288c65978a9b68)), closes [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783) [#783](https://github.com/vonage/vivid/issues/783)
+
+
+
 ## [2.6.2](https://github.com/vonage/vivid/compare/v2.6.1...v2.6.2) (2021-05-05)